First draft of typeref-reified generics
A lot more work to be done: Currently this technics doesn't work. But we
have all the needed information to reify any type and build schema
accordingly.

Also it opens way to variable-arity generics which may be useful for cli
usage (method like `data` can easily be introduced, which creates as big
tuple as needed)

As of now I can't understand if we really can reify generics in
following simple case:

```kotlin
data class<A>(val first: A, val second: A)
```

Here we can easily see that we can't just positionally search for
reified parameter types because second parameter doen't exist.

Signed-off-by: asm0dey <pavel.finkelshtein@gmail.com>