Fixes #16
Detailed research revealed that we're incorrectly inferring field names
if they don't confirm to javaBean specification. It turns out that we're
inferring field name from getter which is not always correct.

So in this commit we're adding new wrapper KStructField, which contains
information not only on field name, but also information on getter name
which gives us abiity to filter data not by inferred field name, but
by getter name which is much more safe - there is no way kotlin will
allow the scapitalized and non capitalized properties to co-exist in
same class, like `Country` and `country`, so usage of getter looks more
correct way to handle this situation.

Signed-off-by: Pasha Finkelshteyn <asm0dey@jetbrains.com>