feat: adds support for Tuple encoding
Up to this moment, there was no ability to work with `Tuple`s in Kotlin API for Apache Spark, which stopped us from

1. Mixing Scala and Kotlin code in one project
2. Call some operations like `select` returning typed tuples

Also, potentially it could bring unavoidable performance hits when we're forcing users to use explicit Tuple → data class conversions. Costs should be negligible, but we can't really measure it and, consequentially, should give users a choice of Kotlin idiomatic way or potentially more performant code,

We thank @Jolanrensen for their commitment to the project and the huge effort to fix this issue. Thank you very much!